two tests in place and functioning, some bug fixes and cleanup done
