+++ src/ComboDisplayDiv.tsx
@@ -1,13 +1,13 @@
 import styled from "styled-components"
 import { LinearProgress } from "@mui/material";
 import { useContext, useEffect, useRef, useState } from "react";
-import { LineClearContext } from "./App";
+import { LineClearContext } from "./context";
 
 const ComboTimeDiv = styled.div `
-position: absolute;
 border: 1px solid #000000;
 height: 30%;
 width: 65%;
+margin-left: -5%;
 `
 
 const ContainerDiv = styled.div `
@@ -26,26 +26,17 @@ border-radius: 50%;
 `
 
 type ComboProps = {
-    circleRef: React.RefObject<HTMLDivElement | null>;
-    circleWidth: number;
     combo: number;
     setCombo: React.Dispatch<React.SetStateAction<number>>
     score: number;
     setScore: React.Dispatch<React.SetStateAction<number>>
 }
 
-export const ComboDisplayDiv = ({circleRef, circleWidth, combo, setCombo, score, setScore}: ComboProps) => {
+export const ComboDisplayDiv = ({combo, setCombo, score, setScore}: ComboProps) => {
     const progressRef = useRef<HTMLDivElement>(null);
     const lineClearBus = useContext(LineClearContext);
 
     const [comboTimer, setComboTimer] = useState(0);
-
-    useEffect(() => {
-        if(progressRef.current) {
-            const barMargin = circleWidth + 10;
-            progressRef.current.style.marginLeft =  `${barMargin}px` ;
-        }
-    }, [circleRef, circleWidth]);
 
     useEffect(() => { 
         const comboUp = (linesCleared: number) => {
@@ -58,8 +49,6 @@ export const ComboDisplayDiv = ({circleRef, circleWidth, combo, setCombo, score,
             //timeout defers process to the next tick so this doesn't try to run mid render of Game
             setTimeout(() => {setCombo(combo+linesCleared)}, 0);
             setTimeout(() => {setScore(score+100*(1+(combo*0.1))*linesCleared)}, 0);
-            console.log({score, combo, linesCleared});
-            console.log("Added score: ", 100*(1+(combo*0.1))*linesCleared);
         }
 
         let comboIntervalId = -1;
@@ -91,13 +80,13 @@ export const ComboDisplayDiv = ({circleRef, circleWidth, combo, setCombo, score,
     }, [comboTimer]);
 
     return (
-        <ContainerDiv>
+        <ContainerDiv data-testid="comboDisp">
+            <CircleDiv>
+                <p>{String(combo)}</p>
+            </CircleDiv>
             <ComboTimeDiv ref={progressRef}>
                 <LinearProgress variant="determinate" value={comboTimer} sx={{ height: '100%', "& .MuiLinearProgress-bar": {transition: 10}} }/>
             </ComboTimeDiv>
-            <CircleDiv ref={circleRef}>
-                <p>{String(combo)}</p>
-            </CircleDiv>
         </ContainerDiv>
     );
 }
