+++ src/Game.tsx
@@ -7,7 +7,8 @@ import { canFall } from "./canFall"
 import { Shape, shapeChart } from "./constants"
 import { outerOffsets } from "./outerOffsets"
 import { UpNextCanvas } from "./UpNextCanvas"
-import { LineClearContext, PauseContext } from "./App"
+import { LineClearContext } from "./context"
+import { PauseContext } from "./context"
 import { GameOverModal } from "./GameOverModal"
 import { StartMenuModal } from "./StartMenuModal"
